moved the user get feelings into an index.js component file in dashboards components folder.
