+++ client/src/App.js
@@ -33,14 +33,14 @@ function App() {
             <Route path="/login" element={<Login />} />
             <Route path="/community" element={<Community />} />
             {/* <Route path="/comments" element={<Feeling />} /> */}
-            {/* <Route path="/dashboard" element={<Dashboard />} /> */}
+            <Route path="/dashboard" element={<Dashboard />} />
           </Routes>
           <Routes>
             <Route path="/about" element={<Footer />} />
             <Route path="/login" element={<Footer />} />
             <Route path="/community" element={<Footer />} />
             {/* <Route path='/comments' element={<Footer />} /> */}
-            {/* <Route path='/dashboard' element={<Footer />} /> */}
+            <Route path='/dashboard' element={<Footer />} />
           </Routes>
         </Router>
       </ApolloProvider>
