+++ client/src/pages/Dashboard.js
@@ -1,39 +1,28 @@
 import React, { useState } from 'react';
-import { Link } from 'react-router-dom';
-//import { useQuery } from '@apollo/client';
-//import { QUERY_MATCHUPS } from '../utils/js/queries';
+import { useQuery } from '@apollo/client';
+import { GET_FEELINGS } from '../utils/js/queries';
 import { Col, Row, Button } from 'antd';
-import { PlusCircleFilled, EditFilled, DeleteFilled } from '@ant-design/icons';
+import { PlusCircleFilled } from '@ant-design/icons';
 import { Layout } from 'antd';
 import '../utils/css/Index.css';
 import CreateModal from '../components/dashboard/NewPost';
-import EditModal from '../components/dashboard/EditPost';
+
+import UserFeelings from '../components/dashboard/Index';
 
 const { Sider, Content } = Layout;
 
-const Dashboard = ({ open, onOK, onCancel}) => {
-
+const Dashboard = () => {
+    const { loading, data } = useQuery(GET_FEELINGS);
     const [createPost, setCreatePost] = useState(false);
-    const [editPost, setEditPost] = useState(false);
-
-    // const { loading, data} = useQuery(QUERY_MATCHUPS, {
-    //     fetchPolicy: 'no-cache'
-    // });
-
-    // const matchupList = data?.matchups || [];
 
     const newPostClick = () => {
         setCreatePost(!createPost);
     }
 
-    const editPostClick = () => {
-        setEditPost(!editPost);
-    }
-
     return (
         <main className="mainContent">
             {createPost && ( <CreateModal />)}
-            {editPost && ( <EditModal />)}
+
             <Layout>
                 <Sider>
                     <Col className='D-Col'>
@@ -47,27 +36,11 @@ const Dashboard = ({ open, onOK, onCancel}) => {
                     </Col>
                 </Sider>
                 <Layout>
-                    <Content className="D-Content">
-                        <Row className="D-Content-Row" wrap={true}>
-                            <Row className="D-Content-Row-Post">
-                                <Row className="D-Post-Row">
-                                    <Col className="D-Post-Title">
-                                        <h3>Qu'est-ce que le Lorem Ipsum?</h3>
-                                    </Col>
-                                    <Col className="D-Post-Edit">
-                                        <p className="D-Post-Edit-E"><EditFilled onClick={editPostClick}/></p>
-                                        <p className="D-Post-Edit-D"><DeleteFilled /></p>
-                                    </Col>
-                                </Row>
-                                <Row className="D-Post-Text">
-                                    <p>Le Lorem Ipsum est simplement du faux texte employé dans la composition et la mise en page avant impression. Le Lorem Ipsum est le faux texte standard de l'imprimerie depuis les années 1500, quand un imprimeur anonyme assembla ensemble des morceaux de texte pour réaliser un livre spécimen de polices de texte.</p>
-                                </Row>
-                                <Row className="D-Post-Date">
-                                    <p>Nov 29th, 2022</p>
-                                </Row>
-                            </Row>
-                        </Row>
-                    </Content>
+                    {loading ? (
+                        <div>Loading...</div>
+                    ) : (
+                        <UserFeelings />
+                    )}
                 </Layout>
             </Layout>
         </main>
